Feature/customers

+++ app/Http/Controllers/ProductController.php
@@ -3,68 +3,76 @@
 namespace App\Http\Controllers;
 
 use App\Models\Product;
+use App\Utils\FileHandler;
 use Illuminate\Http\Request;
 use App\Models\ProductCategory;
 use App\Http\Requests\ProductRequest;
+use Illuminate\Foundation\Auth\Access\AuthorizesRequests;
 
 class ProductController extends Controller
 {
-    public function index()
-    {
-        $products = Product::mostRecent()
-            ->with('category')
-            ->with('user')
-            ->simplePaginate(8);
-        $categories = ProductCategory::all();
+    use AuthorizesRequests;
 
+    /**
+     * Display a listing of products with pagination and categories.
+     */
+    public function index(?Request $request)
+    {
+        $filters = $request->only([
+            'category_id', 'name', 'stock'
+        ]);
         return inertia('Product/Index', [
-            'products' => $products,
-            'categories' => $categories,
+            'filters' => $filters,
+            'products' => Product::mostRecent()
+                ->filter($filters)
+                ->with(['category', 'user'])
+                ->simplePaginate(4)
+                ->withQueryString(),
+            'categories' => ProductCategory::all(),
         ]);
     }
 
+    /**
+     * Store a newly created product in the database.
+     */
     public function store(ProductRequest $request)
     {
-        // Valider les données
         $validated = $request->validated();
+        $validated['filename'] = FileHandler::handleImageUpload($request, 'filename', 'products/product.jpg');
 
-        // Vérifier et gérer le fichier téléversé (cover_photo)
-        if ($request->hasFile('image')) {
-            $filePath = $request->file('image')->store('products', 'public');
-            $validated['filename'] = $filePath;
-        }
-        // Créer le produit avec les données validées et l'utilisateur associé
         $product = $request->user()->products()->create($validated);
 
-        $product->filename = $filePath;
+        $product->filename = $validated['filename'];
         $product->save();
 
-        // Rediriger avec un message de succès
-        return redirect()->back()->with('success', 'Produit créé avec succès.');
+        return redirect()->route('product.index')->with('success', 'Product created successfully.');
     }
 
+    /**
+     * Update the specified product in the database.
+     */
     public function update(ProductRequest $request, Product $product)
     {
-        // Valider les données
+        $this->authorize('update', $product);
+
         $validated = $request->validated();
-
-        // Vérifier et gérer le fichier téléversé (image)
-        if ($request->hasFile('image')) {
-            $filePath = $request->file('image')->store('products', 'public');
-            $validated['filename'] = $filePath;
-        }
-
-        // Mettre à jour le produit avec les données validées
+        $validated['filename'] = FileHandler::handleImageUpload($request, 'filename', 'products/product.jpg', $product->filename);
+        $product->filename = $validated['filename'];
         $product->update($validated);
 
-        // Rediriger avec un message de succès
-        return redirect()->back()->with('success', 'Produit mis à jour avec succès.');
+        return redirect()->route('product.index')->with('success', 'Product updated successfully.');
     }
 
+    /**
+     * Remove the specified product from the database.
+     */
     public function destroy(Product $product)
     {
+        $this->authorize('delete', $product);
+
+        FileHandler::deleteFile($product->filename, 'products/product.jpg');
         $product->delete();
 
-        return redirect()->back()->with('success', 'Product deleted successfully.');
+        return redirect()->route('product.index')->with('success', 'Product deleted successfully.');
     }
 }
